Move explorer in another widget
